feat: full application refactoring — security, performance, UI/UX, features
Constraint 1 — Security Hardening:
- Encrypted Hive boxes with AES-256 via iOS Keychain/Android KeyStore
- TLS 1.3 enforcement + certificate pinning in DioClient
- Input sanitization utility (HTML/JS/SQL injection prevention)
- PINs moved from Hive to platform secure enclave

Constraint 2 — Performance & Memory (4-6GB RAM target):
- 50MB/50-image cache cap in main.dart
- BackgroundProcessor for isolate-based heavy computation
- PaginatedListView with shimmer skeleton loading
- Paginated session retrieval in LocalDataSource

Constraint 3 — UI/UX Modernization:
- useMaterial3: true on both themes
- Full light theme with system theme switching
- AppSpacing/AppRadius/AppSizes standardized constants
- Dynamic text scaling (0.8x-1.6x clamp)
- AccessibilityWrapper with 48dp min touch targets + Semantics

Constraint 4 — Feature Integration:
- KanbanProvider: task tracker with todo/in_progress/done states
- KanbanScreen: 3-column board with drag-and-drop + swipe-delete
- SyncService: offline-first queue with connectivity-triggered sync
- CalendarService: device calendar API integration
- Async secure PIN loading from enclave

Also fixes pre-existing issues:
- score_calculator_service.dart num→double type errors
- enterprise_test.dart null→DateTime in JournalModel tests
- focus_session_screen.dart unused variable warning

+++ lib/core/services/score_calculator_service.dart
@@ -91,12 +91,15 @@ class ScoreCalculatorService {
   }) {
     return {
       'base': 100,
-      'social_media_penalty': -(socialMediaOverGoalMinutes * 0.8).clamp(0, 35),
-      'screen_time_penalty': -(screenTimeOverGoalMinutes * 0.3).clamp(0, 20),
-      'override_penalty': -(overrideTaps * 3.0).clamp(0, 15),
-      'incomplete_sessions_penalty': -(incompleteSessions * 5.0).clamp(0, 15),
+      'social_media_penalty':
+          -(socialMediaOverGoalMinutes * 0.8).clamp(0, 35).toDouble(),
+      'screen_time_penalty':
+          -(screenTimeOverGoalMinutes * 0.3).clamp(0, 20).toDouble(),
+      'override_penalty': -(overrideTaps * 3.0).clamp(0, 15).toDouble(),
+      'incomplete_sessions_penalty':
+          -(incompleteSessions * 5.0).clamp(0, 15).toDouble(),
       'no_morning_session_penalty': noSessionBy11am ? -5.0 : 0,
-      'habits_penalty': -(habitsNotCompleted * 3.0).clamp(0, 15),
+      'habits_penalty': -(habitsNotCompleted * 3.0).clamp(0, 15).toDouble(),
       'poor_sleep_penalty': poorSleep ? -10.0 : 0,
       'focus_sessions_bonus': (completedSessions * 8.0).clamp(0, 40),
       'apps_beat_goal_bonus': (appsBeatGoal * 5.0).clamp(0, 15),
